split out async zmq code
